Say what the player answered when a document will not parse
Both of these turned up watching a player change input.

The status poll fails there occasionally with "returned XML this crate could
not read" and nothing more: the parse failure was kept as a source that no
caller printed, and the document that caused it is gone by the time anyone
looks. Error::Xml now carries the head of the response.

`indexing` was typed as u8. It is not a flag - the player reports how many
songs it has indexed so far, which reaches five figures on any real library,
and a value over 255 would fail to parse and take the whole status document
down with it.

+++ crates/bluos/src/status.rs
@@ -273,8 +273,10 @@ pub struct Status {
     pub prid: Option<u32>,
     /// Service identity.
     pub sid: Option<u32>,
-    /// Non-zero while the player is indexing a share.
-    pub indexing: Option<u8>,
+    /// How many songs the player has indexed so far, and zero when it is not
+    /// indexing. A library runs to five figures, so this is not a small number
+    /// and not a percentage — there is no total to measure it against.
+    pub indexing: Option<u32>,
     /// Minutes left on the sleep timer; empty when it is off.
     pub sleep: Option<String>,
     /// Mirrors `/SyncStatus`'s etag, so a change in grouping is visible from a
